pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "recast"
version = "0.1.0"
description = "Radio broadcast to podcast converter with local AI"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
dependencies = [
    "ffmpeg-python>=0.2.0",
    "faster-whisper>=1.1.0",
    "ollama>=0.3.0",
    "sqlite-utils>=3.35",
    "watchdog>=4.0",
    "fastapi>=0.110",
    "uvicorn>=0.29",
    "feedgen>=1.0",
    "typer>=0.12",
    "structlog>=24.0",
    "mutagen>=1.47",
    "websockets>=12.0",
    "python-multipart>=0.0.9",
]

[project.optional-dependencies]
gpu = [
    "pyannote.audio>=3.0",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
    "pytest-cov>=5.0",
]

[project.scripts]
recast = "recast.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["recast"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
